requirements.txt
streamlit==1.40.0
langchain==0.3.7
langchain-mistralai==0.2.4
langgraph==0.2.53
langchain-core==0.3.19
python-dotenv==1.0.1
fastapi==0.115.5
uvicorn[standard]==0.32.1
requests==2.32.3
pydantic==2.10.2
